requirements.txt
streamlit
numpy
pandas
plotly
altair
python-dotenv
requests
websockets
httpx
deepgram-sdk
